gan_training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.5.0
tensorboard>=2.8.0
albumentations>=1.3.0
scikit-image>=0.19.0
tqdm>=4.64.0
wandb>=0.13.0
lpips>=0.1.4
pytorch-fid>=0.3.0
